refactor(rr): simplify RR into single RRStep class, remove batch machinery
- Merge RecursiveReflector into RRStep (RecursiveAgent subclass, no wrapper)
- Delete runner.py — all RR logic now in ace/steps/rr_step.py
- Remove 12 batch methods and batch-specific prompt formatting
- Move sandbox, timeout, max_output_chars to AgenticConfig/AgenticDeps
- Simplify prompts to remove batch format placeholders
- RecursiveConfig now only overrides max_output_chars
- RRDeps now only adds trace_data and skillbook_text
