+++ ace/implementations/rr/__init__.py
@@ -1,15 +1,13 @@
 """Recursive Reflector implementation.
 
-Config, prompts, tools, and core logic live here. The pipeline step
-definition lives in ``ace.steps.rr_step``.
+Config, prompts, and tools live here. The ``RecursiveReflector`` class
+and the pipeline step ``RRStep`` live in ``ace.steps.rr_step``.
 """
 
 from .config import RecursiveConfig as RRConfig
-from .runner import RecursiveReflector
 from .tools import RRDeps
 
 __all__ = [
     "RRConfig",
     "RRDeps",
-    "RecursiveReflector",
 ]
