+++ ace/implementations/rr/config.py
@@ -7,15 +7,10 @@ from ...core.recursive_agent import AgenticConfig
 
 @dataclass
 class RecursiveConfig(AgenticConfig):
-    """Configuration for the RecursiveReflector.
+    """Configuration for the Recursive Reflector.
 
-    Extends :class:`AgenticConfig` with RR-specific fields for sandbox
-    execution and output limits.
-
-    Attributes:
-        timeout: Timeout in seconds for each code execution (default: 30.0)
-        max_output_chars: Maximum characters per code execution output before truncation (default: 20000)
+    Inherits all fields from :class:`AgenticConfig`. Overrides
+    ``max_output_chars`` for larger trace outputs.
     """
 
-    timeout: float = 30.0
-    max_output_chars: int = 20_000
+    max_output_chars: int = 50_000
