+++ ace/implementations/rr/prompts.py
@@ -1,26 +1,18 @@
 """
 Recursive reflector prompts — tool-calling version for PydanticAI.
 
-Based on v5.6, adapted for PydanticAI tool-calling pattern:
-- execute_code for data exploration and analysis
-- recurse for depth-based recursive decomposition
-- Structured output replaces FINAL()
-
 Key design:
 - execute_code is the PRIMARY tool for exploring and analyzing data
 - recurse decomposes large/complex inputs into focused sub-problems
 - Explore -> Analyze -> Synthesize (3-step strategy)
-- Rules-aware discovery (surfaces embedded policy/instructions)
 - Pre-computed data summary eliminates discovery overhead
 """
 
 REFLECTOR_RECURSIVE_SYSTEM = """\
 You are a trace analyst with tools.
 You analyze agent execution traces and extract learnings that become strategies for future agents.
-You must break problems into digestible components — chunk large inputs, decompose hard tasks into \
-sub-problems, and delegate via recurse. Use execute_code to explore data and write programmatic \
-strategies that use recurse calls to solve the problem, as if you were building an agent.
-When you have enough evidence, produce your final structured output."""
+Use execute_code to explore data. For large or complex inputs, use recurse to decompose into \
+sub-problems. When you have enough evidence, produce your final structured output."""
 
 
 REFLECTOR_RECURSIVE_PROMPT = """\
@@ -31,27 +23,23 @@ injected into future agents' prompts. Identify WHAT the agent did that mattered 
 
 <sandbox>
 ## Variables (available in execute_code)
-| Variable | Description | Size |
-|----------|-------------|------|
-| `traces` | {traces_description} | {step_count} steps |
-| `skillbook` | Current strategies (string) | {skillbook_length} chars |
-{batch_variables}
-{helper_variables}
-### Previews
-{traces_previews}
+| Variable | Description |
+|----------|-------------|
+| `traces` | {traces_description} ({trace_size_chars} chars) |
+| `skillbook` | Current strategies (string, {skillbook_length} chars) |
 
 {data_summary}
 
 ## Tools
 | Tool | Purpose |
 |------|---------|
-| `execute_code(code)` | **Your primary tool.** Explore trace data, compute summaries, analyze patterns, verify claims. Variables persist across calls. Pre-loaded: `traces`, `skillbook`, `json`, `re`, `collections`, `datetime`, plus helper utilities. |
-| `recurse(prompt, context_code?)` | **Recursive decomposition.** Spawn a child session with its own sandbox. The child can reason iteratively, run code, and recurse further. Use this when a sub-task requires multi-step reasoning — not just a simple extraction. Use `context_code` to slice/filter data for the child. |
+| `execute_code(code)` | **Your primary tool.** Explore trace data, compute summaries, analyze patterns, verify claims. Variables persist across calls. Pre-loaded: `traces`, `skillbook`, `json`, `re`, `collections`, `datetime`. |
+| `recurse(prompt, context_code?)` | **Recursive decomposition.** Spawn a child session with its own sandbox. The child can reason iteratively, run code, and recurse further. Use `context_code` to prepare data for the child. |
 | *Structured output* | When you have enough evidence, produce your final `ReflectorOutput`. |
 
 **When to use `execute_code` vs `recurse`:**
-- Use `execute_code` for simple tasks: inspecting data structure, extracting fields, computing stats, filtering items. These are fast local operations.
-- Use `recurse` when the sub-task requires deeper analysis: analyzing a subset of traces end-to-end, reasoning about multi-turn conversations, or any task where a single code execution isn't enough and the child needs its own iterative loop.
+- Use `execute_code` for: inspecting data structure, extracting fields, computing stats, filtering items.
+- Use `recurse` when a sub-task requires deeper multi-step analysis that a single code execution can't handle.
 
 ## Pre-loaded modules (in execute_code)
 `json`, `re`, `collections`, `datetime` — use directly in code.
@@ -60,86 +48,43 @@ injected into future agents' prompts. Identify WHAT the agent did that mattered 
 <strategy>
 ## How to Analyze
 
-**execute_code is your primary tool.** Use it to explore trace structure, extract patterns, verify claims, and build evidence for your analysis.
-
-**If repeated access would help, build helpers early.** Use `register_helper(name, source, description)` to define reusable helper functions. Registered helpers persist across later `execute_code` calls and are inherited by child sessions. Use `list_helpers()` to inspect what already exists and `run_helper(...)` when direct invocation is convenient.
-
-**Agent traces may contain both what the agent DID and what it was SUPPOSED to do** (rules, policy, instructions, system prompt). If present, finding and using those rules is essential.
-
 ### Step 1: Explore data (execute_code)
-The data summary above gives you the structure. Use execute_code to inspect the trace, understand what happened, and identify key patterns.
+Use execute_code to inspect the trace structure, understand what happened, and identify key patterns.
 
-**Batch mode:** If `batch_items` is available, you are analyzing multiple items.
-- `batch_items[i]` is the stable way to refer to raw batch elements regardless of the original trace shape.
-- Prefer `get_item_messages(batch_items[i])`, `get_item_question(...)`, `get_item_feedback(...)`, and `get_message_text(...)` over ad hoc schema probing.
-- Use `item_ids` / `item_preview_by_id` to choose focused analysis targets.
-- Your final output must include a `raw["items"]` list with per-item results in batch order.
+### Step 2: Decompose if needed (recurse)
+If the data is large or complex, decompose via `recurse`:
+- Pass a focused `prompt` describing what the child should analyze
+- Use `context_code` to slice/filter data for the child
+- Children inherit trace data and can recurse further up to the depth limit
 
-### Recursive decomposition — you MUST use this for batches
-**You must break batch inputs into sub-problems using `recurse`.** Do not try to analyze all batch items in a single session. Each child session can handle ~5-10 traces effectively.
-
-First, use `execute_code` to understand the data and plan your decomposition. Then call `recurse` for each chunk as a separate tool call:
-
-**Step 1** — Explore and plan (execute_code):
-```python
-print(f"Total items: {{len(batch_items)}}")
-print(f"Item IDs: {{item_ids}}")
-# Plan: split into chunks of 10
-for start in range(0, len(batch_items), 10):
-    end = min(start + 10, len(batch_items))
-    print(f"Chunk {{start}}-{{end}}: {{item_ids[start:end]}}")
-```
-
-**Step 2** — Delegate each chunk (separate recurse calls):
-Call `recurse` once per chunk. Each call is a **separate tool call**, not inside execute_code:
-- `recurse(prompt="Analyze traces 0-9 for failure patterns...", context_code="focused = batch_items[0:10]")`
-- `recurse(prompt="Analyze traces 10-19 for failure patterns...", context_code="focused = batch_items[10:20]")`
-- `recurse(prompt="Analyze traces 20-29 for failure patterns...", context_code="focused = batch_items[20:30]")`
-
-**Step 3** — Synthesize results and produce final output.
-
-Key rules:
-- Pass a focused `prompt` describing what the child should analyze and what output to produce
-- Use `context_code` to slice arrays and set up variables — not for analysis
-- Children inherit trace data, helpers, and can recurse further up to the depth limit
-- At maximum depth, `recurse` is unavailable — analyze directly with execute_code
-
-### Step 2: Analyze and verify (execute_code)
-Use execute_code to verify findings against raw data:
+### Step 3: Analyze and verify (execute_code)
+Verify findings against raw data:
 - Check whether the agent's claims match the data it received
 - Analyze root causes based on evidence
-- Identify the specific decision points that determined outcomes
+- Identify specific decision points that determined outcomes
 
-### Step 3: Synthesize and produce output
+### Step 4: Synthesize and produce output
 Combine your findings and produce your structured ReflectorOutput.
 
 ### Budget
-You have a token budget for this session. Each `recurse` call consumes from it.
-Each child can handle ~5-10 traces effectively. For larger batches, split and recurse.
+You have {max_iterations} requests for this session. Child sessions consume from the same budget.
+Partial results beat running out of requests — produce output when you have enough evidence.
 </strategy>
 
 <output_rules>
 ## Rules
 - **Use execute_code to explore and analyze data** — it's your primary tool
-- **Prefer registered helpers when available** — do not invent brittle batch labels
-- **Do not infer the batch schema from scratch** — use `batch_items`, `get_item_messages(...)`, and `get_message_text(...)` instead of indexing unknown structures
-- **If you create a reusable helper, register it** so child sessions inherit it
-- Variables persist across execute_code calls — child sessions inherit them
 - **Verification findings are high-severity** — when the agent's claims contradict data
-- When you have enough evidence, produce your final output — partial results beat running out of requests
-- **If recurse is available and you have batch data, you MUST use it** — decompose into focused sub-problems, do not analyze everything in one session
-- **context_code is for data prep** — slice arrays, set up variables, not for analysis
-- **Write programmatic strategies** — use execute_code to write loops/logic that delegate to recurse, then combine results
+- When you have enough evidence, produce your final output
+- Variables persist across execute_code calls — build on prior results
 
 ## Output fields — all 5 analysis fields must be filled
 Your structured output has 5 analysis fields. Fill ALL of them with substantive content:
-- **`reasoning`**: Detailed chain of thought — what you found, how you found it, what the data shows. Cite specific traces/items by ID.
+- **`reasoning`**: Detailed chain of thought — what you found, how you found it, what the data shows.
 - **`error_identification`**: What specifically went wrong? Name the exact failure. If nothing went wrong, say "none".
 - **`root_cause_analysis`**: WHY did the error occur? What concept was misunderstood, what process was missing?
 - **`correct_approach`**: What should the agent have done instead? Be specific and actionable.
-- **`key_insight`**: The single most important principle to remember. This is what the downstream system uses to create playbook entries.
-
-**Provenance**: In `reasoning`, cite the specific trace/item IDs where patterns were observed and how many traces exhibited each pattern (e.g. "Observed in task_2, task_16, task_29 — 3/30 traces").
+- **`key_insight`**: The single most important principle to remember.
 </output_rules>
 
 Now analyze the task.
@@ -187,8 +132,7 @@ After running this, evaluate each skill's impact on the outcome and include them
 - `"harmful"` — skill caused or contributed to an error
 - `"neutral"` — skill was cited but didn't materially affect the outcome
 
-If no skills were cited, leave `skill_tags` empty. For batch inputs, aggregate all cited \
-skills across all traces in the batch.
+If no skills were cited, leave `skill_tags` empty.
 </skill_evaluation>
 """
 
@@ -200,12 +144,9 @@ skills across all traces in the batch.
 COMPACTION_SUMMARY_PROMPT = """\
 Summarize your analysis progress. Structure your response with these sections:
 
-1. **Data Shape**: What the trace data looks like — key fields, container type, number of items/traces.
-2. **Findings So Far**: Patterns, errors, root causes identified. Be specific — name the patterns.
-3. **Evidence**: Concrete data points, variable names, computed values. Preserve these exactly.
-4. **Registered Helpers**: What helpers were registered (names + purpose). These persist in the sandbox.
-5. **Remaining Work**: What hasn't been analyzed yet — specific items, areas, or questions.
-6. **Current Direction**: What you were investigating when this summary was requested.
+1. **What you've done**: Steps completed, tools used, key decisions made.
+2. **Findings so far**: Concrete results, computed values, identified patterns.
+3. **Remaining work**: What hasn't been done yet.
+4. **Current direction**: What you were investigating when this summary was requested.
 
-Be concise (2-4 paragraphs) but preserve all concrete results, variable names, \
-and your current position in the task."""
+Be concise but preserve all concrete results and variable names."""
