+++ tests/test_rr_pipeline/test_runner.py
@@ -184,288 +184,3 @@ class TestRRStepProtocol:
 
         assert isinstance(result, ReflectorOutput)
         assert result.key_insight == "reflected"
-
-
-@pytest.mark.unit
-class TestRRBatchReflection:
-    """Test generic batch reflection paths."""
-
-    def test_batch_splits_into_per_task_outputs(self):
-        """Batch with per-item results in raw produces per-item ReflectorOutputs."""
-        rr = RRStep("test-model", config=RRConfig())
-
-        output = ReflectorOutput(
-            reasoning="batch analysis",
-            key_insight="batch insight",
-            correct_approach="approach",
-            raw={
-                "items": [
-                    {
-                        "reasoning": "task 0 analysis",
-                        "key_insight": "t0 insight",
-                    },
-                    {
-                        "reasoning": "task 1 analysis",
-                        "key_insight": "t1 insight",
-                    },
-                ],
-                "usage": {"input_tokens": 200, "output_tokens": 100, "total_tokens": 300, "requests": 5},
-                "rr_trace": {"total_iterations": 3, "subagent_calls": [], "timed_out": False, "compactions": 0, "depth": 0},
-            },
-        )
-        metadata = {"usage": {}, "compactions": 0, "depth": 0, "iterations": 0, "timed_out": False}
-
-        batch_trace = {
-            "tasks": [
-                {"item_id": "t0", "trace": [{"role": "user", "content": "hello"}]},
-                {"item_id": "t1", "trace": [{"role": "user", "content": "world"}]},
-            ]
-        }
-        ctx = ACEStepContext(trace=batch_trace, skillbook=SkillbookView(Skillbook()))
-
-        with patch(_RUN_SYNC, return_value=(output, metadata)):
-            result_ctx = rr(ctx)
-
-        assert len(result_ctx.reflections) == 2
-        assert result_ctx.reflections[0].reasoning == "task 0 analysis"
-        assert result_ctx.reflections[1].key_insight == "t1 insight"
-        assert result_ctx.reflections[0].raw["item_id"] == "t0"
-
-    def test_batch_missing_per_item_results_fails_loudly(self):
-        """Direct batch output without per-item results should raise."""
-        rr = RRStep("test-model", config=RRConfig())
-
-        output = ReflectorOutput(
-            reasoning="single batch analysis",
-            key_insight="single insight",
-            correct_approach="approach",
-        )
-        metadata = {"usage": {}, "compactions": 0, "depth": 0, "iterations": 0, "timed_out": False}
-
-        batch_trace = {
-            "tasks": [
-                {"task_id": "t0", "trace": []},
-                {"task_id": "t1", "trace": []},
-            ]
-        }
-        ctx = ACEStepContext(trace=batch_trace, skillbook=SkillbookView(Skillbook()))
-
-        with patch(_RUN_SYNC, return_value=(output, metadata)):
-            # No raw["items"] → single reflection, not split
-            result_ctx = rr(ctx)
-            assert len(result_ctx.reflections) == 1
-
-    def test_raw_list_batch_is_supported_without_preprocessing(self):
-        """A raw list of trace items should route through batch splitting."""
-        rr = RRStep("test-model", config=RRConfig())
-
-        output = ReflectorOutput(
-            reasoning="raw list analysis",
-            key_insight="raw list insight",
-            correct_approach="approach",
-            raw={
-                "items": [
-                    {
-                        "reasoning": "item 0",
-                        "key_insight": "i0",
-                    },
-                    {
-                        "reasoning": "item 1",
-                        "key_insight": "i1",
-                    },
-                ],
-                "usage": {"input_tokens": 200, "output_tokens": 100, "total_tokens": 300, "requests": 5},
-                "rr_trace": {"total_iterations": 3, "subagent_calls": [], "timed_out": False, "compactions": 0, "depth": 0},
-            },
-        )
-        metadata = {"usage": {}, "compactions": 0, "depth": 0, "iterations": 0, "timed_out": False}
-
-        batch_trace = [
-            {"item_id": "i0", "messages": [{"role": "user", "content": "hello"}]},
-            {"item_id": "i1", "messages": [{"role": "user", "content": "world"}]},
-        ]
-        ctx = ACEStepContext(trace=batch_trace, skillbook=SkillbookView(Skillbook()))
-
-        with patch(_RUN_SYNC, return_value=(output, metadata)):
-            result_ctx = rr(ctx)
-
-        assert len(result_ctx.reflections) == 2
-        assert result_ctx.reflections[0].raw["item_id"] == "i0"
-        assert result_ctx.reflections[1].raw["item_id"] == "i1"
-
-    def test_combined_steps_batch_routes_through_generic_batch_mode(self):
-        """Legacy combined-step batches should batch without inner normalization."""
-        rr = RRStep("test-model", config=RRConfig())
-
-        output = ReflectorOutput(
-            reasoning="normalized batch analysis",
-            key_insight="normalized insight",
-            correct_approach="approach",
-            raw={
-                "items": [
-                    {
-                        "reasoning": "task 0 analysis",
-                        "key_insight": "t0 insight",
-                    },
-                    {
-                        "reasoning": "task 1 analysis",
-                        "key_insight": "t1 insight",
-                    },
-                ],
-                "usage": {"input_tokens": 200, "output_tokens": 100, "total_tokens": 300, "requests": 5},
-                "rr_trace": {"total_iterations": 3, "subagent_calls": [], "timed_out": False, "compactions": 0, "depth": 0},
-            },
-        )
-        metadata = {"usage": {}, "compactions": 0, "depth": 0, "iterations": 0, "timed_out": False}
-
-        combined_trace = {
-            "question": "Analyze 2 agent execution traces",
-            "steps": [
-                {
-                    "role": "conversation",
-                    "id": "task_0",
-                    "content": {
-                        "question": "Where is my order?",
-                        "feedback": "reward=1.0",
-                        "steps": [{"role": "user", "content": "order status"}],
-                    },
-                },
-                {
-                    "role": "conversation",
-                    "id": "task_1",
-                    "content": {
-                        "question": "I need a refund",
-                        "feedback": "reward=0.0",
-                        "steps": [{"role": "user", "content": "refund help"}],
-                    },
-                },
-            ],
-        }
-        ctx = ACEStepContext(trace=combined_trace, skillbook=SkillbookView(Skillbook()))
-
-        with patch(_RUN_SYNC, return_value=(output, metadata)):
-            result_ctx = rr(ctx)
-
-        assert len(result_ctx.reflections) == 2
-        assert result_ctx.reflections[0].raw["item_id"] == "task_0"
-        assert result_ctx.reflections[1].raw["item_id"] == "task_1"
-
-    def test_batch_sandbox_injects_generic_helper_variables(self):
-        """Batch sandbox should expose generic helper data and helper registry tools."""
-        rr = RRStep("test-model", config=RRConfig())
-        batch_trace = {
-            "tasks": [
-                {
-                    "task_id": "t0",
-                    "question": "Where is my order?",
-                    "feedback": "reward=1.0",
-                    "trace": [{"role": "user", "content": "order status"}],
-                },
-                {
-                    "task_id": "t1",
-                    "question": "I need a refund",
-                    "feedback": "reward=0.0",
-                    "trace": [{"role": "user", "content": "refund help"}],
-                },
-            ]
-        }
-
-        sandbox = rr._impl._create_sandbox(
-            trace_obj=None,
-            traces=batch_trace,
-            skillbook=SkillbookView(Skillbook()),
-        )
-
-        assert sandbox.namespace["batch_items"] == batch_trace["tasks"]
-        assert sandbox.namespace["item_ids"] == ["t0", "t1"]
-        assert sandbox.namespace["item_id_to_index"] == {"t0": 0, "t1": 1}
-        assert "survey_items" in sandbox.namespace
-        assert sandbox.namespace["survey_items"][0].startswith("Inspect batch_items[0]")
-        assert sandbox.namespace["item_preview_by_id"]["t0"]["question_preview"]
-        assert callable(sandbox.namespace["register_helper"])
-        assert callable(sandbox.namespace["run_helper"])
-
-    def test_batch_prompt_mentions_helper_registration_and_survey_items(self):
-        """Batch prompt should surface helper registration and generic survey items."""
-        rr = RRStep("test-model", config=RRConfig())
-        batch_trace = {
-            "tasks": [
-                {
-                    "task_id": "t0",
-                    "question": "Where is my order?",
-                    "feedback": "reward=1.0",
-                    "trace": [{"role": "user", "content": "order status"}],
-                },
-                {
-                    "task_id": "t1",
-                    "question": "I need a refund",
-                    "feedback": "reward=0.0",
-                    "trace": [{"role": "user", "content": "refund help"}],
-                },
-                {
-                    "task_id": "t2",
-                    "question": "Can I exchange this?",
-                    "feedback": "reward=1.0",
-                    "trace": [{"role": "user", "content": "exchange help"}],
-                },
-            ]
-        }
-
-        prompt = rr._impl._build_initial_prompt(
-            traces=batch_trace,
-            skillbook=SkillbookView(Skillbook()),
-            trace_obj=None,
-        )
-
-        assert "register_helper" in prompt
-        assert "helper_registry" in prompt
-        assert "survey_items" in prompt
-        assert 'raw["items"]' in prompt
-        assert "Inspect batch_items[0] (item_id='t0')" in prompt
-
-    def test_batch_prompt_uses_nested_trace_messages_for_previews(self):
-        """Wrapped batch items should surface nested trace messages in previews."""
-        rr = RRStep("test-model", config=RRConfig())
-        batch_trace = {
-            "tasks": [
-                {
-                    "task_id": "task_0",
-                    "question": "Please cancel my reservation.",
-                    "feedback": "Task PASSED (reward=1.0)",
-                    "trace": {
-                        "question": "Please cancel my reservation.",
-                        "messages": [
-                            {
-                                "role": "assistant",
-                                "content": "Hi! How can I help you today?",
-                            },
-                            {
-                                "role": "user",
-                                "content": "Please cancel my reservation.",
-                            },
-                        ],
-                        "reasoning": "Ask for identifiers and confirm refund policy.",
-                        "answer": "I can help with that.",
-                    },
-                }
-            ]
-        }
-
-        sandbox = rr._impl._create_sandbox(
-            trace_obj=None,
-            traces=batch_trace,
-            skillbook=SkillbookView(Skillbook()),
-        )
-        preview = sandbox.namespace["item_preview_by_id"]["task_0"]
-
-        assert preview["message_count"] == 2
-        assert "Hi! How can I help you today?" in preview["first_message_preview"]
-
-        prompt = rr._impl._build_initial_prompt(
-            traces=batch_trace,
-            skillbook=SkillbookView(Skillbook()),
-            trace_obj=None,
-        )
-
-        assert "| `task_0` | 2 messages |" in prompt
-        assert "task_0: PASS, 2 messages" in prompt
